+++ dashboard-widgets.el
@@ -1,6 +1,6 @@
 ;;; dashboard-widgets.el --- A startup screen extracted from Spacemacs  -*- lexical-binding: t -*-
 
-;; Copyright (c) 2016-2022 emacs-dashboard maintainers
+;; Copyright (c) 2016-2023 emacs-dashboard maintainers
 ;;
 ;; Author     : Rakan Al-Hneiti <rakan.alhneiti@gmail.com>
 ;; Maintainer : Jesús Martínez <jesusmartinez93@gmail.com>
@@ -166,11 +166,11 @@ preserved."
 
 (defcustom dashboard-navigator-buttons nil
   "Specify the navigator buttons.
-The format is: 'icon title help action face prefix suffix'.
+The format is: `icon title help action face prefix suffix`.
 
 Example:
-'((\"☆\" \"Star\" \"Show stars\" (lambda (&rest _)
-                                    (show-stars)) 'warning \"[\" \"]\"))"
+`((\"☆\" \"Star\" \"Show stars\" (lambda (&rest _)
+                                    (show-stars)) warning \"[\" \"]\"))"
   :type '(repeat (repeat (list string string string function symbol string string)))
   :group 'dashboard)
 
@@ -215,15 +215,16 @@ Default value is `official', it displays the Emacs logo.  `logo' displays Emacs
 alternative logo.  An integer value is the index of text banner.  A string
 value must be a path to a .PNG or .TXT file.  If the value is nil then no banner
 is displayed."
-  :type '(choice (const  :tag "offical"   official)
-                 (const  :tag "logo"      logo)
-                 (const  :tag "phdemacs"  phdemacs)
-                 (string :tag "a png or txt path"))
-  :group 'dashboard)
-
-(defcustom dashboard-buffer-last-width nil
-  "Previous width of dashboard-buffer."
-  :type  'integer
+:type '(choice   (const   :tag "no banner" nil)
+                 (const   :tag "offical"   official)
+                 (const   :tag "logo"      logo)
+                 (const   :tag "phdemacs"  phdemacs)
+                 (integer :tag "index of a text banner")
+                 (string  :tag "a path to an image or text banner")
+                 (cons    :tag "an image and text banner"
+                          (string :tag "image banner path")
+                          (string :tag "text banner path"))
+                 )
   :group 'dashboard)
 
 (defcustom dashboard-item-generators
@@ -399,7 +400,7 @@ If nil it is disabled.  Possible values for list-type are:
 
 (defun dashboard-str-len (str)
   "Calculate STR in pixel width."
-  (let ((width (window-font-width))
+  (let ((width (frame-char-width))
         (len (dashboard-string-pixel-width str)))
     (+ (/ len width)
        (if (zerop (% len width)) 0 1))))  ; add one if exceeed
@@ -504,34 +505,85 @@ If MESSAGEBUF is not nil then MSG is also written in message buffer."
     (overlay-put ov 'face 'dashboard-heading))
   (when shortcut (insert (format " (%s)" shortcut))))
 
-(defun dashboard-center-line (string)
-  "Center a STRING accoring to it's size."
-  (insert (make-string (max 0 (floor (/ (- dashboard-banner-length
-                                           (+ (string-width string) 1)) 2))) ?\ )))
+(defun dashboard-center-text (start end)
+  "Center the text between START and END."
+  (save-excursion
+    (goto-char start)
+    (let ((width 0))
+      (while (< (point) end)
+        (let* ((line-str (buffer-substring (line-beginning-position) (line-end-position)))
+               (line-length (dashboard-str-len line-str)))
+          (setq width (max width line-length)))
+        (forward-line 1))
+      (let ((prefix (propertize " " 'display `(space . (:align-to (- center ,(/ width 2)))))))
+        (add-text-properties start end `(line-prefix ,prefix indent-prefix ,prefix))))))
+
+(defun dashboard-insert-center (&rest strings)
+  "Insert STRINGS in the center of the buffer."
+  (let ((start (point)))
+    (apply #'insert strings)
+    (dashboard-center-text start (point))))
 
 ;;
 ;; BANNER
 ;;
-(defun dashboard-insert-ascii-banner-centered (file)
-  "Insert banner from FILE."
-  (let ((ascii-banner
-         (with-temp-buffer
-           (insert-file-contents file)
-           (let ((banner-width 0))
-             (while (not (eobp))
-               (let ((line-length (- (line-end-position) (line-beginning-position))))
-                 (if (< banner-width line-length)
-                     (setq banner-width line-length)))
-               (forward-line 1))
-             (goto-char 0)
-             (let ((margin
-                    (max 0 (floor (/ (- dashboard-banner-length banner-width) 2)))))
-               (while (not (eobp))
-                 (insert (make-string margin ?\ ))
-                 (forward-line 1))))
-           (buffer-string))))
-    (put-text-property 0 (length ascii-banner) 'face 'dashboard-text-banner ascii-banner)
-    (insert ascii-banner)))
+
+(defun dashboard-get-banner-path (index)
+  "Return the full path to banner with index INDEX."
+  (concat dashboard-banners-directory (format "%d.txt" index)))
+
+(defun dashboard--image-supported-p (img)
+  "Return non-nil if IMG exists and is a supported image type."
+  ;; In Emacs 29.1 we could use `image-supported-file-p'. However:
+  ;; - We need to support Emacs 26.
+  ;; - That function will only look at filenames, this one will inspect the file data itself.
+  (and (file-exists-p img) (ignore-errors (image-type-available-p (image-type img)))))
+
+(defun dashboard-choose-banner ()
+  "Return a plist specifying the chosen banner based on `dashboard-startup-banner'."
+  (pcase dashboard-startup-banner
+    ('nil nil)
+    ('official
+     (append (when (image-type-available-p 'png)
+               (list :image dashboard-banner-official-png))
+             (list :text (dashboard-get-banner-path 1))))
+    ('logo
+     (append (when (image-type-available-p 'png)
+               (list :image dashboard-banner-logo-png))
+             (list :text (dashboard-get-banner-path 1))))
+    ('phd-emacs
+     (append (when (image-type-available-p 'png)
+               (list :image dashboard-banner-phdemacs-png))
+             (list :text (dashboard-get-banner-path 3))))
+    ((pred integerp)
+     (list :text (dashboard-get-banner-path dashboard-startup-banner)))
+    ((pred stringp)
+     (pcase dashboard-startup-banner
+       ((pred (lambda (f) (not (file-exists-p f))))
+        (message "could not find banner %s, use default instead" dashboard-startup-banner)
+        (list :text (dashboard-get-banner-path 1)))
+       ((pred (string-suffix-p ".txt"))
+        (list :text (if (file-exists-p dashboard-startup-banner)
+                        dashboard-startup-banner
+                      (message "could not find banner %s, use default instead" dashboard-startup-banner)
+                      (dashboard-get-banner-path 1))))
+       ((pred dashboard--image-supported-p)
+        (list :image dashboard-startup-banner
+              :text (dashboard-get-banner-path 1)))
+       (_
+        (message "unsupported file type %s" (file-name-nondirectory dashboard-startup-banner))
+        (list :text (dashboard-get-banner-path 1)))))
+    (`(,img . ,txt)
+     (list :image (if (dashboard--image-supported-p img)
+                      img
+                    (message "could not find banner %s, use default instead" img)
+                    dashboard-banner-official-png)
+           :text (if (and (file-exists-p txt) (string-suffix-p ".txt" txt))
+                     txt
+                   (message "could not find banner %s, use default instead" txt)
+                   (dashboard-get-banner-path 1))))
+    (_
+     (message "unsupported banner config %s" dashboard-startup-banner))))
 
 (defun dashboard--type-is-gif-p (image-path)
   "Return if image is a gif.
@@ -539,39 +591,71 @@ String -> bool.
 Argument IMAGE-PATH path to the image."
   (eq 'gif (image-type image-path)))
 
-(defun dashboard-insert-image-banner (banner)
-  "Display an image BANNER."
-  (when (file-exists-p banner)
-    (let* ((title dashboard-banner-logo-title)
-           (size-props
-            (append (when (> dashboard-image-banner-max-width 0)
-                      (list :max-width dashboard-image-banner-max-width))
-                    (when (> dashboard-image-banner-max-height 0)
-                      (list :max-height dashboard-image-banner-max-height))))
-           (spec
-            (cond ((dashboard--type-is-gif-p banner)
-                   (create-image banner))
-                  ((image-type-available-p 'imagemagick)
-                   (apply 'create-image banner 'imagemagick nil size-props))
-                  (t
-                   (apply 'create-image banner nil nil
-                          (when (and (fboundp 'image-transforms-p)
-                                     (memq 'scale (funcall 'image-transforms-p)))
-                            size-props)))))
-           ;; TODO: For some reason, `elisp-lint' is reporting error void
-           ;; function `image-size'.
-           (size (when (fboundp 'image-size) (image-size spec)))
-           (width (car size))
-           (left-margin (max 0 (floor (- dashboard-banner-length width) 2))))
-      (goto-char (point-min))
+(defun dashboard-insert-banner ()
+  "Insert the banner at the top of the dashboard."
+  (goto-char (point-max))
+  (when-let (banner (dashboard-choose-banner))
+    (insert "\n")
+    (let ((start (point))
+          buffer-read-only
+          text-width
+          image-spec)
       (insert "\n")
-      (insert (make-string left-margin ?\ ))
-      (insert-image spec)
-      (when (dashboard--type-is-gif-p banner) (image-animate spec 0 t))
+      ;; If specified, insert a text banner.
+      (when-let (txt (plist-get banner :text))
+        (insert-file-contents txt)
+        (put-text-property (point) (point-max) 'face 'dashboard-text-banner)
+        (setq text-width 0)
+        (while (not (eobp))
+          (let ((line-length (- (line-end-position) (line-beginning-position))))
+            (if (< text-width line-length)
+                (setq text-width line-length)))
+          (forward-line 1)))
+      ;; If specified, insert an image banner. When displayed in a graphical frame, this will
+      ;; replace the text banner.
+      (when-let (img (plist-get banner :image))
+        (let ((size-props
+               (append (when (> dashboard-image-banner-max-width 0)
+                         (list :max-width dashboard-image-banner-max-width))
+                       (when (> dashboard-image-banner-max-height 0)
+                         (list :max-height dashboard-image-banner-max-height)))))
+          (setq image-spec
+                (cond ((dashboard--type-is-gif-p img)
+                       (create-image img))
+                      ((image-type-available-p 'imagemagick)
+                       (apply 'create-image img 'imagemagick nil size-props))
+                      (t
+                       (apply 'create-image img nil nil
+                              (when (and (fboundp 'image-transforms-p)
+                                         (memq 'scale (funcall 'image-transforms-p)))
+                                size-props))))))
+        (add-text-properties start (point) `(display ,image-spec))
+        (when (dashboard--type-is-gif-p img) (image-animate image-spec 0 t)))
+      ;; Finally, center the banner (if any).
+      (when-let* ((text-align-spec `(space . (:align-to (- center ,(/ text-width 2)))))
+                  (image-align-spec `(space . (:align-to (- center (0.5 . ,image-spec)))))
+                  (prop
+                   (cond
+                    ;; Both an image & text banner.
+                    ((and image-spec text-width)
+                     ;; The quoting is intentional. This is a conditional display spec that will
+                     ;; align the banner at redisplay time.
+                     `((when (display-graphic-p) . ,image-align-spec)
+                       (when (not (display-graphic-p)) . ,text-align-spec)))
+                    ;; One or the other.
+                    (text-width text-align-spec)
+                    (image-spec image-align-spec)
+                    ;; No banner.
+                    (t nil)))
+                  (prefix (propertize " " 'display prop)))
+        (add-text-properties start (point) `(line-prefix ,prefix wrap-prefix ,prefix)))
       (insert "\n\n")
-      (when title
-        (dashboard-center-line title)
-        (insert (format "%s\n\n" (propertize title 'face 'dashboard-banner-logo-title)))))))
+      (add-text-properties start (point) '(cursor-intangible t inhibit-isearch t))))
+  (when dashboard-banner-logo-title
+    (dashboard-insert-center (propertize dashboard-banner-logo-title 'face 'dashboard-banner-logo-title))
+    (insert "\n\n"))
+  (dashboard-insert-navigator)
+  (dashboard-insert-init-info))
 
 ;;
 ;; INIT INFO
@@ -582,51 +666,7 @@ Argument IMAGE-PATH path to the image."
     (let ((init-info (if (functionp dashboard-init-info)
                          (funcall dashboard-init-info)
                        dashboard-init-info)))
-      (dashboard-center-line init-info)
-      (insert (propertize init-info 'face 'font-lock-comment-face)))))
-
-(defun dashboard-get-banner-path (index)
-  "Return the full path to banner with index INDEX."
-  (concat dashboard-banners-directory (format "%d.txt" index)))
-
-(defun dashboard-choose-banner ()
-  "Return the full path of a banner based on the dotfile value."
-  (when dashboard-startup-banner
-    (cond ((eq 'official dashboard-startup-banner)
-           (if (and (display-graphic-p) (image-type-available-p 'png))
-               dashboard-banner-official-png
-             (dashboard-get-banner-path 1)))
-          ((eq 'logo dashboard-startup-banner)
-           (if (and (display-graphic-p) (image-type-available-p 'png))
-               dashboard-banner-logo-png
-             (dashboard-get-banner-path 1)))
-          ((eq 'phdemacs dashboard-startup-banner)
-           (if (and (display-graphic-p) (image-type-available-p 'png))
-               dashboard-banner-phdemacs-png
-             (dashboard-get-banner-path 3)))
-          ((integerp dashboard-startup-banner)
-           (dashboard-get-banner-path dashboard-startup-banner))
-          ((stringp dashboard-startup-banner)
-           (if (and (file-exists-p dashboard-startup-banner)
-                    (or (string-suffix-p ".txt" dashboard-startup-banner)
-                        (and (display-graphic-p)
-                             (image-type-available-p (intern (file-name-extension
-                                                              dashboard-startup-banner))))))
-               dashboard-startup-banner
-             (message "could not find banner %s, use default instead" dashboard-startup-banner)
-             (dashboard-get-banner-path 1)))
-          (t (dashboard-get-banner-path 1)))))
-
-(defun dashboard-insert-banner ()
-  "Insert Banner at the top of the dashboard."
-  (goto-char (point-max))
-  (let ((banner (dashboard-choose-banner)) buffer-read-only)
-    (when banner
-      (if (image-type-available-p (intern (file-name-extension banner)))
-          (dashboard-insert-image-banner banner)
-        (dashboard-insert-ascii-banner-centered banner))
-      (dashboard-insert-navigator)
-      (dashboard-insert-init-info))))
+      (dashboard-insert-center (propertize init-info 'face 'font-lock-comment-face)))))
 
 (defun dashboard-insert-navigator ()
   "Insert Navigator of the dashboard."
@@ -661,15 +701,13 @@ Argument IMAGE-PATH path to the image."
                          :button-suffix suffix
                          :format "%[%t%]")
           (insert " ")))
-      (let* ((width (current-column)))
-        (beginning-of-line)
-        (dashboard-center-line (make-string width ?\s))
-        (end-of-line))
+      (dashboard-center-text (line-beginning-position) (line-end-position))
       (insert "\n"))
     (insert "\n")))
 
 (defmacro dashboard-insert-section (section-name list list-size shortcut-id shortcut-char action &rest widget-params)
   "Add a section with SECTION-NAME and LIST of LIST-SIZE items to the dashboard.
+
 SHORTCUT-CHAR is the keyboard shortcut used to access the section.
 ACTION is theaction taken when the user activates the widget button.
 WIDGET-PARAMS are passed to the \"widget-create\" function."
@@ -737,11 +775,11 @@ to widget creation."
   "Insert footer of dashboard."
   (when-let ((footer (and dashboard-set-footer (dashboard-random-footer))))
     (insert "\n")
-    (dashboard-center-line footer)
-    (insert dashboard-footer-icon)
-    (insert " ")
-    (insert (propertize footer 'face 'dashboard-footer))
-    (insert "\n")))
+    (dashboard-insert-center
+     dashboard-footer-icon
+     " "
+     (propertize footer 'face 'dashboard-footer)
+     "\n")))
 
 ;;
 ;; Truncate
@@ -1162,11 +1200,15 @@ Also,it set text properties that latter are used to sort entries and perform dif
          (todo (or (org-get-todo-state) ""))
          (org-level-face (nth (- (org-outline-level) 1) org-level-faces))
          (todo-state (format org-agenda-todo-keyword-format todo)))
-    (when (null (get-text-property 0 'face headline))
-      (add-face-text-property 0 (length headline) org-level-face t headline))
-    (when (null (get-text-property 0 'face todo-state))
-      (add-face-text-property 0 (length todo-state) (org-get-todo-face todo) t todo-state))
+    (dashboard-agenda--set-face org-level-face headline)
+    (dashboard-agenda--set-face (org-get-todo-face todo) todo-state)
     (concat todo-state " " headline)))
+
+(defun dashboard-agenda--set-face (face text)
+  "Add FACE to TEXT but inherit height from `dashboard-items-face'.
+If not height is found on FACE or `dashboard-items-face' use `default'."
+  (let ((height (face-attribute 'dashboard-items-face :height nil 'default)))
+    (add-face-text-property 0 (length text) `((:height ,height) ,face) nil text)))
 
 (defun dashboard-agenda--formatted-time ()
   "Get the scheduled or dead time of an entry.  If no time is found return nil."
@@ -1187,7 +1229,8 @@ point."
   (let ((scheduled-time (org-get-scheduled-time (point)))
         (deadline-time (org-get-deadline-time (point)))
         (entry-timestamp (dashboard-agenda--entry-timestamp (point)))
-        (due-date (dashboard-due-date-for-agenda)))
+        (due-date (dashboard-due-date-for-agenda))
+        (now (current-time)))
     (unless (and (not (org-entry-is-done-p))
                  (not (org-in-archived-heading-p))
                  (or (and scheduled-time
@@ -1195,6 +1238,7 @@ point."
                      (and deadline-time
                           (org-time-less-p deadline-time due-date))
                      (and entry-timestamp
+                          (org-time-less-p now entry-timestamp)
                           (org-time-less-p entry-timestamp due-date))))
       (point))))
 
@@ -1232,14 +1276,16 @@ This is what `org-agenda-exit' do."
 
 (defun dashboard-agenda--sorted-agenda ()
   "Return agenda sorted by time.
-For now, it only works when dashboard-agenda has been filter by time
-and dashboard-agenda-sort is not nil."
+
+For now, it only works when dashboard-agenda has been filter by time and
+dashboard-agenda-sort is not nil."
   (let ((agenda (dashboard-get-agenda))
         (sort-function (dashboard-agenda--sort-function)))
     (sort agenda sort-function)))
 
 (defun dashboard-agenda--sort-function ()
   "Get the function use to sorted the agenda.
+
 Depending on the list `dashboard-agenda-sorting-strategy' use this strategies to
 build a predicate to compare each enty.
 This is similar as `org-entries-lessp' but with a different aproach."
@@ -1247,6 +1293,7 @@ This is similar as `org-entries-lessp' but with a different aproach."
 
 (defun dashboard-agenda--build-sort-function (strategies)
   "Build a predicate to sort the dashboard agenda.
+
 If `STRATEGIES' is nil then sort using the nil predicate.  Look for the strategy
 predicate, the attributes of the entry and compare entries.  If no predicate is
 found for the strategy it uses nil predicate."